+++ packages/world-postgres/src/message.ts
@@ -3,10 +3,10 @@ import * as z from 'zod';
 import { Base64Buffer } from './zod.js';
 
 /**
-/* pgboss is using JSON under the hood, so we need to base64 encode
-/* the body to ensure binary safety
-/* maybe later we can have a `blobs` table for larger payloads
- **/
+ * graphile-worker is using JSON under the hood, so we need to base64 encode
+ * the body to ensure binary safety
+ * maybe later we can have a `blobs` table for larger payloads
+ */
 export const MessageData = z.object({
   attempt: z.number().describe('The attempt number of the message'),
   messageId: MessageId.describe('The unique ID of the message'),