+++ packages/rollup/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workflow/rollup
 
+## 4.0.0-beta.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.1-beta.51
+
+## 4.0.0-beta.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.1-beta.50
+
 ## 4.0.0-beta.15
 
 ### Patch Changes