+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workflow/vite
 
+## 4.0.0-beta.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.1-beta.51
+
+## 4.0.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.1-beta.50
+
 ## 4.0.0-beta.8
 
 ### Patch Changes